Switch to PSR2
The PEAR coding standard was written for PHP 4, only moderately
adjusted for PHP 5, and never really was updated for PHP 5.3 or even
later versions – in a strict sense it isn't even compatible with PHP 7.

Therefore we switch to PSR2 and fix all errors and most warnings. We
also get rid of most of the doc block comments, which were required by
the PEAR CS, but actually do more harm than good (at least nowadays).