Close #19: Change data storage format
To allow us to maintain Occupancy and HourlyOccupancy independent from
stored data, we switch to JSON, which offers also the possibility to
edit the data files manually.

We do the necessary migration transparently on demand, so users won't
have to run the migration explicitly, what saves us having to
prominently document this step, but still getting support requests.