Close #10: Add more states
Instead of hard-coding the maximum state to 3, we make it configurable
and support an arbitrary amount of states. Users have to manually add
respective state labels and colors, though.

Also, we slightly change the logic of the list views to show entries
for all states except the zero state, regardless whether they are
labeled.