No HTML in language strings
Let's keep it simple for end-users who don't want to mess around with
HTML directly. This way all language strings can also be used in plain
text, and above all it's rather clear that all language string will
have to be properly escaped in an HTML context (we implement that a bit
later, though).