Fix #35 (Dots in page URL break calendar navigation)
The "page URLs" of CMSimple_XH are very special; they are usually not
processed by the usual PHP features, and as such are not subject to the
PHP URL parameter name replacements (spaces and periods are replaced by
underscores).  Thus, we must not process them with `http_build_query()`,
but rather refer to them via `$su`.

To avoid doing this in `CalendarController::modifyUrl()`, thereby
coupling the unrelated URL handling more to the controller, we extract
a simple `Url` class which hopefully does the job properly.

A nice side-effect of this new URL handling is that we automatically
produce more CMSimple_XH-like URLs for the starting page, whose "page
URL" is empty.  This is why we had to update a couple of tests.