Drop CalendarController inheritance
Class inheritance is generally a double-edged sword, but we certainly
do not want to have protected properties.  While we could self-
encapsulate the access to those, there is actually no need for sub-
typing the controllers.  Thus we make CalendarController a trait, so we
can easily re-use its code.  This allows us to get rid of the protected
properties, and also of some protected methods.  Only the abstract
method in the trait need to stay abstract, since abstract private is
only supported as of PHP 8.0.0.