Close #42: Support fractional hour_interval
We only support `1/2`, `1/3`, `1/4`, and `1/6`, which should be
sufficient for almost all use cases, and should not cause too much
trouble when we switch to float arithmetic.

To keep that somewhat simple, we format the times now unconditionally
as `hh:mm` what causes a minor BC break, but is likely an improvement
for most users.  We do not yet support proper internationalization of
that format.

We support transparent migration to the slightly changed storage
format; previously we stored `YYYY-oo-dd-hh`; now we additionally store
the minutes (i.e. `YYYY-oo-dd-hh:mm`).